python/requirements.txt
# FastAPI 서버 의존성
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PostgreSQL & Vector
psycopg2-binary>=2.9.0
pgvector>=0.2.0

# OpenAI
openai>=1.0.0

# 기존 CV 파이프라인 의존성
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
numpy>=1.19.0
pandas>=1.2.0
albumentations>=1.0.0
segmentation-models-pytorch>=0.2.0
Pillow>=8.0.0
tqdm>=4.60.0
shapely>=2.0.0

# RAG 시스템
chromadb>=0.5.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
